--- a/package.json
+++ b/package.json
@@ -25,7 +25,6 @@
 		"lucide-react": "^0.371.0",
 		"next": "14.2.2",
 		"next-themes": "^0.3.0",
-		"npm": "^10.5.2",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"react-hook-form": "^7.51.3",