This PR contains a simple change in the next.config.js and is ignore typing checks and linting check in build, this improves a lot the vercel build time and local build time, see my local test

I have a HP Laptop windows 11, Core i5-1235U, 24gb ram 500gb ssd m.2

Without config improve and whitout .next folder cache: 26sec
![Captura de pantalla 2024-07-18 021453](https://github.com/user-attachments/assets/a8f271d5-e6ba-4d0c-8ebb-20b7627c5865)

With .next cache 11sec:
![Captura de pantalla 2024-07-18 021909](https://github.com/user-attachments/assets/e6e6f535-fa90-4527-80dd-a788986dbb4e)

With config improve and without cache 19sec:
![Captura de pantalla 2024-07-18 021737](https://github.com/user-attachments/assets/b325589e-3564-4dc7-a23a-510685b0a626)

With config improve and with cache 9sec:
![Captura de pantalla 2024-07-18 021800](https://github.com/user-attachments/assets/9a2850c2-7677-422b-91f9-cf472750b164)

7 sec better without cache
2 sec better with cache

So in vercel we will have faster build times